Relicensing under the Mozilla Public License
See `LICENSE`.

The `UNLICENSE` isn't really a valid license in many places, and
attempts to remove restrictions that the MPL also doesn't apply.  The
MPL is a light copyleft license, which is just what I want.
http://web.archive.org/web/20150907225139/https://lists.opensource.org/pipermail/license-review/2012-January/000052.html

As Mozilla recommends in its license policy guide, tests and
documentation shall be licensed under the Creative Commons Zero
License.  This ensures legal rights to re-use of these files.
https://www.mozilla.org/en-US/MPL/license-policy/